Fixed A Bug In The string_copy(char* ,char* ) function
If the initial length of strd was longer(due to garbae value/the pointer was used for some other purpose) than strs the function did not give the desired behaviour
Suppose strd=Master and strs=Yoda the function will give strd=Yodaer which is not the desired result.
This was fixed by putting a null character at the end to makr the end of the strd.
Thus the strd after the Bug fix will be strd=Yoda\0r which for all puropses will act as strd=Yoda.